Refactor InitiatorSubscriptionLogEvent to LogRequest interface:
Includes the polymorphic break out of EthLogEvent,
RunLogEvent, ServiceAgreementLogEvent.
